+++ src/ProtocolTubePurger.php
@@ -4,8 +4,15 @@ declare(strict_types=1);
 
 namespace Zlikavac32\BeanstalkdLib;
 
+use Ds\Sequence;
+use Ds\Set;
+
 interface ProtocolTubePurger
 {
 
-    public function purge(Protocol $protocol, string ...$tubes): void;
+    /**
+     * @param Sequence|string[] $tubes
+     * @param Set|JobState[] $states
+     */
+    public function purge(Protocol $protocol, Sequence $tubes, Set $states): void;
 }